backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.9
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0        # <-- Add this line (required for EmailStr)
sentence-transformers==2.7.0
chromadb==0.5.0
numpy==1.26.4
scikit-learn>=1.4.0
bcrypt==4.0.1
ollama==0.3.3
httpx==0.27.2